Fix generated Debian source packages.
Partially reverts 476b07de3d594c155594bf0535a8827118b95255.

We moved the debian/ directory into a subdirectory to clean up the root
of the repo, but Debian-native packages require that debian/ be at the
root of the repository.

This fixes the generated Debian source packages so that they can
actually be used to rebuild the package (in Debian parlance, the package
FTBFS).