Fix bug where the generated Fluent Bit filters were incorrectly ordered. (#140)
The new test config is a good example.  It has:

    pipeline2:
      processors: [beta, alpha]
    pipeline1:
      processors: [alpha, beta]

The code previously created a flat list of 4 Fluent Bit filters:

    [pipeline2-beta, pipeline2-alpha, pipeline1-alpha, pipeline1-beta]

which was then sorted, putting the pipeline2 filters out-of-order:

    [pipeline1-alpha, pipeline1-beta, pipeline2-alpha, pipeline2-beta]

This PR keeps the pipelines in groups:

    [[pipeline2-beta, pipeline2-alpha], [pipeline1-alpha, pipeline1-beta]]

and then does not sort within each group:

    [[pipeline1-alpha, pipeline1-beta], [pipeline2-beta, pipeline2-alpha]]

thus maintaining the correct order of filters in the generated Fluent
Bit config.