Initial configuration for multi-distro testing. (#242)
* Initial configuration for multi-distro testing.

The idea is to turn on lots of distros by default (i.e. for new applications), but for the existing applications, to disable everything but debian-10. This allows us to support multi-distro testing going forward without breaking the existing tests. We know there are many issues to sort out before we can really enable the existing tests.

* lt -> lts fix

* reorder platforms