Copy over tests and testing libraries from Google's code silo. (#332)
* Copy over tests and testing libraries from Google's code silo.

These files are not yet picked up by Kokoro. They run, mostly
successfully, with the following commands (PROJECT needs to be replaced
with some valid project):

PROJECT=asdf \
  GOOGLE_APPLICATION_CREDENTIALS="${HOME}/credentials.json" \
  ZONE=us-central1-b \
  PLATFORMS=debian-10 \
  AGENTS_TO_TEST=ops-agent \
  go test -v ops_agent_test.go \
  -test.parallel=1000 \
  -tags=integration_test \
  -timeout=4h

PROJECT=asdf \
  GOOGLE_APPLICATION_CREDENTIALS="${HOME}/credentials.json" \
  ZONE=us-central1-b \
  PLATFORMS=debian-10 \
  SCRIPTS_DIR=third_party_apps_data \
  AGENTS_TO_TEST=ops-agent \
  go test -v third_party_apps_test.go \
  -test.parallel=1000 \
  -tags=integration_test \
  -timeout=4h

This PR is intended to be publicly visible, and should be close to its
final form, but it's not intended to be merged yet. First I will make
Kokoro able to pick up the files and run with those instead of pulling
them from Google's code silo. Once that is working, we can consider
merging this PR or something like it.

This PR also includes various changes to logging_test.sh to fix various
failures uncovered by the CI on the GitHub repo.