apps: genericAccessLogParser for nginx_access (#388)
* apps: genericAccessLogParser for nginx_access

The new genericAccessLogParser component works fine for nginx_access and
does not change the output, so switch to using that for
nginx_access component as well.

* apps: Adjust comments to mention nginx

The comments were pulled from the original apache component, so the
comments only mentioned apache. Add mentions of nginx and nginx docs to
comments.