fluent_forward: Use hash in tag prefix to avoid collissions
The receiver_id and pipeline_id currently allows "." to be used. This
creates a problem as thats what the Lua filter must rely on as a
delimeter when adding the old tag to the logname and removing the
pipeline_id from it. Additionally, allowing "." in the receiver_id
creates issues for when a receiver_id contains another one with a ".".

Consider the following example:
```
logging:
  receivers:
    fluent_logs:
      type: fluent_forward
      listen_host: 1.2.3.4
      listen_port: 5678
    fluent_logs.:
      type: fluent_forward
      listen_host: 1.2.3.4
      listen_port: 5679
  service:
    pipelines:
      fluent_pipeline:
        receivers: [fluent_logs, fluent_logs.]
```

Notice that the receiver_id `fluent_logs.` completely contains the other
receiver_id `fluent_logs`. And so when we try to match on records
ingested by `fluent_logs`, since it will be suffixed by the tag of the
forwarder, we will need to match on `fluent_logs.*`. However because of
this, we will match on records from `fluent_logs.` as well. Which is not
desired. To get around this issue, we prefix the tag with a unique hash
that will deterministically ID the pipeline + receiver combination.

This change does the following:
- Add a hash string as the prefix to the tag to prevent collisions
- Update lua script to only append existing tag to the LogName
- Have the Lua filter update logname for `fluent_forward` by appending
  the existing tag.

Signed-off-by: Ridwan Sharif <ridwanmsharif@google.com>