confgenerator: move modify filter before the processors
This change moves the modify filter to before the processors.
This is useful so the processors can reference the LogName moving
foward. This will be needed for `modify_fields` and `exclude_logs`.

This change also updates the `parser` component to preserve existing
fields on the record when parsing. We do this so the LogName is not
blown away when the parser is used.

Signed-off-by: Ridwan Sharif <ridwanmsharif@google.com>