Only run tests relevant to current PR. (#396)
Only run tests relevant to current PR.

Creates a mode of testing for go test -short where we run a smaller slice of tests.

This is intended to be used with presubmits. Postsubmits should omit the -short flag and run the full suite.

Determines what apps are impacted through extracting app name from the file path of modified files
and checking against the known list of apps available for test.