Unnest an unnecessary ops-agent directory (#484)
* Unnest an unnecessary ops-agent directory

integration_test_data/agent/ops-agent/* is now integration_test_data/agent/*

The extra "ops-agent" level of nesting is a holdover from a time when this data directory held scripts and whatnot for multiple agents at once. But we can simplify things a bit because that is no longer the case.