Refactor to use maps
Changed the approach so that `run()` only has to deal with maps of
metrics keyed on type (created a type `expectedMetricsMap` for this),
and then just changed the helper methods to accommodate.
`writeExpectedMetrics` also writes alphabetically by type now, and
fortunately the existing YAML files are already alphabetical so they
don't also need to change.