Also skip "sqlserver" tests on non-SQL windows
This is to prepare for https://github.com/GoogleCloudPlatform/ops-agent/pull/542, which adds a new application called sqlserver. I think we want to skip that application on vanilla windows, like windows-2019, right?

PS this PR maybe should be part of https://github.com/GoogleCloudPlatform/ops-agent/pull/542, but separately like this is fine too.