Set DEBIAN_FRONTEND=noninteractive for all app install scripts
This avoids prompts that hang for user input, now and in the future.

Currently we have this setting sprinkled around where needed, but I just wasted an hour diagnosing why rabbitmq on ubuntu-minimal hangs, and it's because it's missing this setting. Setting the env variable for all installations will fix this issue and others that will happen in the future.