integration_test: package common -> metadata
The pattern of having a `common` package with a bunch of random
functions thrown in is largely undesirable from a code quality
standpoint. All the code presently in `common` however was related to
metadata, so I renamed the package to `metadata` so no one will be
tempted to throw unrelated code in there.

While I was at it, I also removed the requirement for the integration_test build tag, since it wasn't really necessary for this code.