Refactor the `common` package and `integration_test` tags (#768)
* integration_test: package common -> metadata

The pattern of having a `common` package with a bunch of random
functions thrown in is largely undesirable from a code quality
standpoint. All the code presently in `common` however was related to
metadata, so I renamed the package to `metadata` so no one will be
tempted to throw unrelated code in there.

While I was at it, I also removed the requirement for the integration_test build tag, since it wasn't really necessary for this code.

* gce: remove integration_test tag from GCE package

It also was unnecessary for the integration_test tag to be applied to
this package. It really is only important for the tag to be on the test
files.

* generate_expected_metrics: refactor init

The init function was being called erroneously, which would fail when
the environment does not contain Cloud Monitoring credentials. Instead
of an init, refactor it into a function that does the same thing and is
called at the beginning of the program.

* gce: just kidding it's not worth it, restore tag