Internal: fix build_package.ps1 for release builds
For some reason, `KOKORO_GOB_COMMIT_URL_unified_agents` isn't set for release builds when run from github (instead it's  just `KOKORO_GOB_COMMIT_URL`), I guess because I removed all but one SCM from the job config? Anyway, looking at where the files are physically located is a more direct way to get to the actual files and doesn't expose us to weird decisions on the part of the kokoro team about which variables to set.

This is resurrecting an earlier draft of this code. Corey suggested removing some of the duplication in this version, but i can't think of any simple way to do it. Suggestions welcome