Testing: (noop) remove unnecessary ** on artifacts (#771)
This change is coppying a change made within Google's internal codebase.

The Kokoro docs say that these paths are relative to KOKORO_ARTIFACTS_DIR already and we don't need the `**` prefix. When we have the prefix on there, it can pick up unexpected directories at any level of nesting called `log`.