Internal: add kokoro config for debian 9 "stretch" (#779)
This will kind of fix our release process, which is expecting stretch.gcl to exist at this location. Of course, since stretch is deprecated, the test will still fail, but at least it won't fail with a missing config file error. If we decide to keep running our tests with stretch even though it's deprecated, we will need this file.