Rename rockylinux8.gcl to centos8.gcl (#990)
Our release process uses the name `centos8` even though that is deprecated and what it really runs under the hood is rocky-linux-8. 

I recently noticed that our release process was not running the third party apps test for "centos8" releases, so this PR is getting ready to turn that on.