files: use DB.locking config option to improve input performance
Fluent Bit has this option available in the `tail` input plugin.
From looking at the upstream code, it looks like should help
to reduce the number of syscalls on every commit to the DB
but at the price of locking the access to the database file to
third party programs. But for us, that trade-off seems okay?

We should soak tesk/performance test this to see if we get any visible
benefit.

Signed-off-by: Ridwan Sharif <ridwanmsharif@google.com>