Enable instrumentation scope labels for workload.googleapis.com (#1083)
This PR enables the `instrumentation_library_labels` setting in the Cloud Monitoring exporter for all metrics except the `agent.googleapis.com/` metrics. This will cause `instrumentation_source` and `instrumentation_version` to be set to the values of `InstrumentationScope.Name` and `InstrumentationScope.Version` fields in OpenTelemetry metrics data. For built-in third party receivers, we set the scope name to `agent.googleapis.com/${receiver_name}` and the version to a metrics version number.