Add windows_event_log receiver v2 (#1086)
* Add support for windows_event_log v2

For backwards compatibility, none of the new constraints on `channels`
apply to v1. If we do relax this in the future then many of the
constraints can be replaced with ordinary validator tags bundled with
the validator package.

Integration tests and XML support coming in a separate commit.

* Add integration tests

* Set Use_ANSI on Windows 2012

Avoids the issue where event log fields come
back completely blank.

* Add check for too-new channels on v1 receivers

* Redirect stdout to the event log

Stdout is lost when running as a service, so we
never see any logs from `log.Print`. This change
redirects stdout to the event log instead.

* Add integration test for too-new channels on v1

* Add more unit tests

* Add feature tracking for receiver_version