tools: bring back yamlfmt and run it
This PR restores yamlfmt in the precommit tools since the bug that
caused us to remove it in the first place is fixed. This PR also adds a
tools directory so we can more effectively manage the versions of the
tools. There's some other tools that should be added here probably,
similar to the opentelemetry-operations-collector repo.