Don't use inotify
Ad-hoc testing indicates that throughput improves significantly if we
disable inotify. Let's set up a proper soak test that measures not just
throughput but CPU and RAM as well.

This commit will get us a build, and the next commit will customize the
soak test presubmit to use that build at 100 MB/s.