Fix ARM integration tests (#1309)
* Fix TestPrometheusMetricsWithJSONExporter on ARM

* Fix cassandra on ARM

The apt thing is weird and I'm sad about it. It would be
_nice_ to be able to keep everything in a single apt
command. I'm open to alternative-but-"acceptable" ways
of forcing cassandra to use Java 11 when installed
concurrently with Java 17.

* (DO NOT SUBMIT) Don't skip any 3P tests

* Skip couchbase on ARM

Debian 11 is not supported on ARM.

* Skip oracledb on ARM

Missed a platform change, but we can skip the tests anyway since
they only run on centos_rhel and we're just doing Debian 11 for
now.

* Skip mongodb3.6 on ARM

* Update mongodb tests to 6.0

5.0 has had enough soak time and 6.0 is the oldest version
available for ARM distros, so it should be fine to bump
the version now.

* Fix mongodb on ARM

* Skip MySQL on Debian 11 ARM

* Add mysql_general support for MariaDB

* Copy MySQL tests to MariaDB

No changes yet, just making a copy.

* Add tests for MariaDB on ARM

* Simplify mariadb scripts to make sure they actually test defaults

* Update format to properly handle multi-line MariaDB general logs

* Add TODO for more MariaDB testing

* Use MongoDB 6.0 on RedHat too

* mongodb-org-shell has been replaced by mongodb-mongosh, with a different version numbering scheme

* Configure MariaDB to use an error log file instead of journald

* Handle MariaDB slow query log entries with and without times

* Use constants for all of the various MySQL and MariaDB time formats, to make sure none are missed

* Add Debian MariaDB error log path, and update comment to indicate that the other default log path is the upstream default

* subsystem and errorCode are not always present

* Update value regxes to match both MySQL and MariaDB

* Implement the MariaDB + old MySQL slow query log format

* Fix reporting of connected database

* Document MariaDB-specific log fields

* Connect to a database for the test query

* Fix log-slow-verbosity value to not have a space (docs say spaces are allowed :( )

* Update goldens

* Fix incorrect matching on missing lines

* Always wrap the value regex in a non-capturing group so that | works correctly

* Use MariaDB versioning in metadata.yaml

I couldn't think of a better way to represent a range of versions so I just lumped it into a single string...

* Add bogus severity field to mysql_general

* mariadb tests use mysql for instrumentation_source

mariadb is defined as a separate 3p test, but it uses the same
integration as mysql so we need to special-case the expected
value for instrumentation_source.

* Make the secondary mariadb instance the replication source, so the primary instance has replication metrics

* Fix speling mistake

* Try running mariadb tests on all Debian and Ubuntu distros

* Fix cassandra JRE dependency

cassandra depends on either `openjdk-11-jre-headless` or
`java11-runtime`, which is provided by `openjdk-11-jre`, not
`openjdk-11-jdk`.

* Enable couchbase on Debian 11

Only on AMD64 because it's still not supported on ARM64.

* apt-key add is deprecated

* Fix wrong keyfile extensions

dearmor'd key streams are binary and need a .gpg file extension.

* Unpin couchbase version on DEB distros

* Restore the sleep for mariadb exercise script

We saw a few flakes (e.g. `ERROR 2002` and mysql_slow missing)
that are possibly explained by the removed sleep.

* Simplify ARCH shell stuff

* Make couchbase install script arch-agnostic

* Temporarily disable mariadb tests on Debian 10 and Ubuntu 20.04 since they require a different config file location

* Skip mariadb's mysql.row_operations on debian-12

* Typo

* Remove extra whitespace

* Disable other DEB platforms

These can be resolved in a follow-up PR; goal for now is just to get
debian-11-arm64 passing.

* Add configuration note about log_error

* Revert row_operations skip on debian-12

debian-12 is being skipped entirely due to logging being broken
so we don't need this for now.

* Remove temporary defaultPlatforms change

We have a full passing test suite now on Debian 11 ARM, so we don't
need to change defaultPlatforms to force tests anymore.

* Remove unused parameter

Meant to do this as part of 478d040.

---------

Co-authored-by: Quentin Smith <quentin@google.com>