prometheus: untyped metrics are preserved in the GMP compatible way (#1384)
Untyped metrics are double written to GMP when using the Ops Agent. This
is similar to what the GMP binary does on GKE. Untyped metrics used to
be written as a gauge (with a `/gauge` suffix) and with this change, it
will be written as a gauge (with the `/unknown` suffix instead) and a
counter (with the `/unknown:counter`) suffix.

This is a bug fix because rate operations and cumulative interpretations
of these metrics would not work before. Some existing users may need to
update their MQL queries, if they relied on the `/gauge` suffix when
querying this metric. PromQL users will not be affected.