Testing: Remove TODO to go back to using old URL (#1411)
Today (as opposed to when I wrote the comment) I think it would be better to avoid the redirector and just use artifact registry directly. This is because the redirector gives us no benefits, and adding a new vendored repo is trickier if you have to involve the redirector. I'm hoping to simplify our documentation about these repos just a bit.