Template assumes `aarch64` for `.Arch`
`.Arch` comes directly from the user-provided env variable into the
helper script, which I'd prefer to be `aarch64` for the sake of
being the more common ARM counterpart to `x86_64`.

That being said, all of the ARM configs here are already suffixed with
`_arm64`, so it makes sense for the generated files to still contain
`arm64`, hence the conversion inside the template.