Add support for BMS machines (#1327)
This commit adds support for a new resource type when Ops Agent is running on BMS machines (detected via environment variables provided to Ops Agent). It also provides a framework for additional resource types/overrides in the future.

The Prometheus receiver does not yet work correctly with the BMS resource type.