Always update log record after exclusion (#1449)
The inner logic for an exclusion filter relies on the creation of
temporary `__match_` fields. Those fields must be removed before the
filter terminates. The existing lua code was returning `0, 0, 0` in the
event that a log doesn't match the exclusion filter, which means "don't
modify the record." But because we were also removing the `__match_`
fields inside the same lua function, those removals are not being
committed back to the log record afterwards.

Simple fix is to return `2, 0, record` instead.