Temporary patch for incorrect test user key file ownership
When we create test VMs, we include an SSH pubkey for a test user as metadata.
The guest agent sees this and creates a system user (with a
`~/.ssh/authorized_keys` file holding the key we defined) accordingly. On some
DLVM images, the `~/.ssh/authorized_keys` file subsequently gets its ownership
attributes corrupted during startup. This prevents us from being able to SSH
into the VM for the remainder of the test.

This change deploys a workaround that manually patches over the corruption at
VM startup.