[confgenerator] Create `MergeInternalLoggingProcessor` in `FilesMixin` to merge processor `MultilineRules`. (#2025)
This PR adds a new `InternalLoggingProcessorMerger` interface that receivers can implement to merge in processors, patterned after the `MetricsProcessorMerger` interface that already exists on the metrics side.

This allows the clean up of the previous hack to merge `ParseMultiline` by munging the generated fluent-bit config, adds support for merging `LoggingProcessorParseMultilineRegex`, and allows 3P app receivers that use multiline parsing to be refactored into macros.

Finally, this PR also cleans up a fluent-bit bug, in which multiline parsers used in `filter` components strip trailing newlines but the same multiline parser used in an `input` component does not.

Co-authored-by: Quentin Smith <quentin@google.com>