Preserve newlines in Invoke-Program output
When you send the output of a native command to a variable, you actually get an array of strings, one for each line. By default, when this array is formatted as a string, the elements are joined by a space which makes the output hard to read. This change pre-formats the cmd output as a string by explicitly joining the elements with newlines instead of spaces.