transformation_test_update again
So this time it removed the same entry it added before. I'm scared now
that this is a new flake and it basically sometimes works sometimes
doesn't. I'm not sure how to handle it if that's the case, but at the
moment I can't figure out at all why this is happening.